server/requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
httpx>=0.23.0
pydantic>=2.0
python-dotenv>=0.19.0
geopy>=2.3.0